+++ apps/teacher/src/androidTest/java/com/instructure/teacher/ui/TeacherToDoDetailsPageTest.kt
@@ -13,7 +13,7 @@
  *     See the License for the specific language governing permissions and
  *     limitations under the License.
  */
-package com.instructure.student.ui.interaction
+package com.instructure.teacher.ui
 
 import android.app.Activity
 import androidx.compose.ui.test.hasText
@@ -22,31 +22,31 @@ import com.instructure.canvas.espresso.common.ToDoDetailsInteractionTest
 import com.instructure.canvas.espresso.mockCanvas.MockCanvas
 import com.instructure.canvas.espresso.mockCanvas.init
 import com.instructure.espresso.InstructureActivityTestRule
-import com.instructure.student.BuildConfig
-import com.instructure.student.activity.LoginActivity
-import com.instructure.student.ui.pages.DashboardPage
-import com.instructure.student.ui.utils.StudentActivityTestRule
-import com.instructure.student.ui.utils.tokenLogin
+import com.instructure.teacher.activities.LoginActivity
+import com.instructure.teacher.BuildConfig
+import com.instructure.teacher.ui.pages.DashboardPage
+import com.instructure.teacher.ui.utils.TeacherActivityTestRule
+import com.instructure.teacher.ui.utils.tokenLogin
 import dagger.hilt.android.testing.HiltAndroidTest
 
 @HiltAndroidTest
-class ToDoDetailsInteractionTest : ToDoDetailsInteractionTest() {
+class TeacherToDoDetailsPageTest : ToDoDetailsInteractionTest() {
+
+    override val activityRule: InstructureActivityTestRule<out Activity>
+            = TeacherActivityTestRule(LoginActivity::class.java)
 
     override val isTesting = BuildConfig.IS_TESTING
-
-    override val activityRule: InstructureActivityTestRule<out Activity> =
-        StudentActivityTestRule(LoginActivity::class.java)
 
     val dashboardPage = DashboardPage()
 
     override fun displaysPageObjects() = Unit
 
     override fun goToToDoDetails(data: MockCanvas) {
-        val student = data.students[0]
-        val token = data.tokenFor(student)!!
-        tokenLogin(data.domain, token, student)
+        val teacher = data.teachers[0]
+        val token = data.tokenFor(teacher)!!
+        tokenLogin(data.domain, token, teacher)
 
-        dashboardPage.clickCalendarTab()
+        dashboardPage.openCalendar()
 
         val todo = data.todos.first()
